backend/requirements.txt
# Web framework
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0

# Repo fetching
gitpython>=3.1

# RAG & embeddings
langchain>=0.3
langchain-community>=0.3
langchain-huggingface>=0.1
chromadb>=0.5
sentence-transformers>=3.1

# LLM (Ollama)
langchain-ollama>=0.2

# Utilities
python-dotenv>=1.0
pydantic>=2.9
sse-starlette>=2.1
httpx>=0.27
